+++ src/sinks/console.rs
@@ -1,9 +1,12 @@
-use super::util::SinkExt;
 use crate::{
     buffers::Acker,
-    event::{self, Event},
+    sinks::util::{
+        encoding::{self, BasicEncoding},
+        SinkExt,
+    },
     topology::config::{DataType, SinkConfig},
 };
+
 use futures::{future, Sink};
 use serde::{Deserialize, Serialize};
 use tokio::{
@@ -29,14 +32,7 @@ impl Default for Target {
 pub struct ConsoleSinkConfig {
     #[serde(default)]
     pub target: Target,
-    pub encoding: Option<Encoding>,
-}
-
-#[derive(Deserialize, Serialize, Debug, Eq, PartialEq, Clone)]
-#[serde(rename_all = "snake_case")]
-pub enum Encoding {
-    Text,
-    Json,
+    pub encoding: Option<BasicEncoding>,
 }
 
 #[typetag::serde(name = "console")]
@@ -52,58 +48,12 @@ impl SinkConfig for ConsoleSinkConfig {
         let sink = FramedWrite::new(output, LinesCodec::new())
             .stream_ack(acker)
             .sink_map_err(|_| ())
-            .with(move |event| encode_event(event, &encoding));
+            .with(move |event| encoding::event_as_string(event, &encoding));
 
         Ok((Box::new(sink), Box::new(future::ok(()))))
     }
 
     fn input_type(&self) -> DataType {
         DataType::Any
-    }
-}
-
-fn encode_event(event: Event, encoding: &Option<Encoding>) -> Result<String, ()> {
-    match event {
-        Event::Log(log) => {
-            if (log.is_structured() && encoding != &Some(Encoding::Text))
-                || encoding == &Some(Encoding::Json)
-            {
-                let bytes = serde_json::to_vec(&log.all_fields())
-                    .map_err(|e| panic!("Error encoding: {}", e))?;
-                String::from_utf8(bytes)
-                    .map_err(|e| panic!("Unable to convert json to utf8: {}", e))
-            } else {
-                let s = log
-                    .get(&event::MESSAGE)
-                    .map(|v| v.to_string_lossy())
-                    .unwrap_or_else(|| "".into());
-                Ok(s)
-            }
-        }
-        Event::Metric(metric) => serde_json::to_string(&metric).map_err(|_| ()),
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::encode_event;
-    use crate::{event::Metric, Event};
-
-    #[test]
-    fn encodes_raw_logs() {
-        let event = Event::from("foo");
-        assert_eq!(Ok("foo".to_string()), encode_event(event, &None));
-    }
-
-    #[test]
-    fn encodes_metrics() {
-        let event = Event::Metric(Metric::Counter {
-            name: "foos".into(),
-            val: 100.0,
-        });
-        assert_eq!(
-            Ok(r#"{"type":"counter","name":"foos","val":100.0}"#.to_string()),
-            encode_event(event, &None)
-        );
     }
 }